requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1.2
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.2
et-xmlfile==1.1.0
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
